mockgcp: update serviceaccount implementation
