+++ mockgcp/mockiam/service.go
@@ -23,33 +23,36 @@ import (
 	"net/http"
 
 	"google.golang.org/grpc"
+	"google.golang.org/protobuf/proto"
 
 	"github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/common"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/common/httpmux"
 	pb "github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/generated/mockgcp/iam/admin/v1"
+	"github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/mockgcpregistry"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/pkg/storage"
 )
+
+func init() {
+	mockgcpregistry.Register(New)
+}
 
 // MockService represents a mocked IAM service.
 type MockService struct {
 	*common.MockEnvironment
 	storage storage.Storage
-
-	serverV1 *ServerV1
 }
 
-type ServerV1 struct {
+type IAMServer struct {
 	*MockService
 	pb.UnimplementedIAMServer
 }
 
 // New creates a MockService
-func New(env *common.MockEnvironment, storage storage.Storage) *MockService {
+func New(env *common.MockEnvironment, storage storage.Storage) mockgcpregistry.MockService {
 	s := &MockService{
 		MockEnvironment: env,
 		storage:         storage,
 	}
-	s.serverV1 = &ServerV1{MockService: s}
 	return s
 }
 
@@ -58,10 +61,19 @@ func (s *MockService) ExpectedHosts() []string {
 }
 
 func (s *MockService) Register(grpcServer *grpc.Server) {
-	pb.RegisterIAMServer(grpcServer, s.serverV1)
+	pb.RegisterIAMServer(grpcServer, &IAMServer{MockService: s})
 }
 
 func (s *MockService) NewHTTPMux(ctx context.Context, conn *grpc.ClientConn) (http.Handler, error) {
-	return httpmux.NewServeMux(ctx, conn, httpmux.Options{},
+	mux, err := httpmux.NewServeMux(ctx, conn, httpmux.Options{},
 		pb.RegisterIAMHandler)
+	if err != nil {
+		return nil, err
+	}
+
+	mux.RewriteHeaders = func(ctx context.Context, response http.ResponseWriter, payload proto.Message) {
+		response.Header().Del("Cache-Control")
+	}
+
+	return mux, nil
 }
